\docType{data}
\name{default_parms}
\alias{default_parms}
\title{Default sodi parameters}
\format{List of 24
 $ K          : num 1168
 $ bbox       : num [1:4] 0 100 0 100
 $ n0         : num 1168
 $ infect0    : num 1
 $ stages0    : num [1:1168] 1 1 1 1 1 1 1 1 1 1 ...
 $ dispersalfn: num 3
 $ seedshadow : num 0
 $ sp_names   : chr [1:3] "Tanoak" "Bay" "Redwood"
 $ sp_stages  : num [1:3] 4 1 1
 $ m          : num [1:6] 0.2 0.2 0.2 0.2 0.2 0.2
 $ seedm      : num [1:6] 1 1 1 1 1 1
 $ times      : num [1:11] 0 10 20 30 40 50 60 70 80 90 ...
 $ f          : num [1:6] 0 0.007 0.2 0.73 0.077 0.019
 $ g          : num [1:6] 0.142 0.2 0.05 0 0 0
 $ d          : num [1:6] 0.006 0.003 0.001 0.032 0.002 0.005
 $ r          : num [1:6] 0.5 0.5 0.5 0.5 0 0
 $ alpha      : num [1:6] 0.005 0.005 0.005 0.005 0 0
 $ lamda      : num [1:6] 1 1 1 1 4 0
 $ beta       : num [1:6] 0.33 0.32 0.3 0.24 0.3 0
 $ mu         : num [1:6] 0.05 0.05 0.05 0.05 0.05 0.05
 $ xi         : num [1:6] 1 1 1 1 1 1
 $ omega      : num [1:6] 1.93 0.76 0.66 1.55 1 1
 $ max_inf    : num [1:6] 10 10 10 10 10 10
 $ beta_meth  : num 2}
\usage{
  default_parms
}
\description{
  Default sodi parameters
}
\keyword{datasets}

