\docType{data}
\name{parms1}
\alias{parms1}
\title{Default sodi parameters}
\format{List of 24
 $ K          : num 500
 $ bbox       : num [1:4] 0 22.4 0 22.4
 $ n0         : num 500
 $ infect0    : num 1
 $ stages0    : num [1:500] 1 1 1 1 1 1 1 1 1 1 ...
 $ dispersalfn: num 3
 $ seedshadow : num 0
 $ sp_names   : chr "Tanoak"
 $ sp_stages  : num 2
 $ m          : num [1:2] 1 1
 $ seedm      : num [1:2] 1 1
 $ times      : num [1:101] 0 1 2 3 4 5 6 7 8 9 ...
 $ f          : num [1:2] 0.01 0.01
 $ g          : num [1:2] 0.1 0
 $ d          : num [1:2] 0.005 0.005
 $ r          : num [1:2] 0.5 0.5
 $ alpha      : num [1:2] 0.005 0.005
 $ lamda      : num [1:2] 0.3 0.3
 $ beta       : num [1:2] 1 1
 $ mu         : num [1:2] 0.05 0.05
 $ xi         : num [1:2] 1 1
 $ omega      : num [1:2] 1 1
 $ max_inf    : num [1:2] 1 1
 $ beta_meth  : num 0}
\usage{
  parms1
}
\description{
  Default sodi parameters
}
\keyword{datasets}

